Create helper class to share some test setup information for HeadersCollectionTests
